Simplify synchronization of config changes
Move the synchronization from the EntriesFile up to the dnsmasqResolver.
Operations which did not modify the contents are ignored (re-add same entry, or
remove a missing entry) so the file is not re-written and reloaded. This also
synchronizes the writing and reloading across the resolve, to ensure that both
config files are fully written when signalling the dnsmasq process to reload.