Clean resolv.conf based on marker comment
Adds a comment "added by docker-resolver" to resolv.conf that is used to clean
up any old lines previously added.  Lines with the comment are stripped on
startup before inserting the new nameserver. This assumes that they were left
by an old process that was killed or otherwise failed to clean up.  On
shutdown, the comment lines are stripped from resolv.conf again.