Extract Docker testing daemon management
Moves the Docker daemon management used for testing out into its own package,
with multiple implementations for using Docker-in-Docker or direct container
management.

Since Docker-in-Docker is not available in CI, this allows running in "native"
mode with a direct connection to the provided Docker host. This is destructive
since it kills all containers running on the Docker host in between test
executions.

Docker management is provided via a "Pool" interface which is implemented
for both "native" mode and Docker-in-Docker. The pool maintains a set of
connections to Docker hosts that can be borrowed and returned. For the "native"
mode, the pool is blocking, and has only one connection since it connects to a
single Docker host. With Docker-in-Docker, new "dind" containers are allocated
as needed, and returned connections can be reused by another test.