Only listen on "udp4" network
Fixes an issue with tests for "Wait()" on CircleCI, which doesn't have IPv6.

During shutdown, the DNS server code sends a query to itself in order to
unblock the socket read loop and continue shutting down. When the connection is
established with the "udp" network, it tries to send to "[::]" which fails in
CircleCI which does not have IPv6. Opening the connection as "udp4" lets it
send to "0.0.0.0" instead, which works as expected.

The resolv.conf support is already limited to IPv4 at this time as well, so
this doesn't impose additional restrictions.